requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.6
djangorestframework==3.12.4
flake8==3.9.2
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
